Return random content
This commit adds a special path for random content:

http://localhost:3068/content/random/coming_soon

This will give you a randomly generated content item, based on the
`coming_soon` schema. We use `govuk_schemas` to do this.

It could be useful to check design with random content, because it
shows where the layout is likely to break with irregular content. It's
also a quick way to find bugs caused by reliance on fields that aren't
required by the schemas.